add missing operations to netlib blas

+++ nd4j-api/src/main/java/org/nd4j/linalg/ops/reduceops/scalarops/Norm2.java
@@ -14,9 +14,9 @@ public class Norm2 extends BaseScalarOp {
 
     @Override
     public double accumulate(INDArray arr, int i, double soFar) {
-        double ret =  soFar + (double) Math.pow(arr.getFloat(i),2);
+        double ret =  soFar +  Math.pow(arr.getDouble(i),2);
         if(i == arr.length() - 1)
-            return (double) Math.sqrt(ret);
+            return  Math.sqrt(ret);
         return ret;
     }
 }
